[Chore #15402620] Resolve eslint issues
